Add files via upload
PDF of a general outline added.. The register /  Login should be the first thing user sees when they open page (if not already logged in).. Advance Search can be either a pop up or a div that drops down over the search term heading and disappears upon search.. or maybe it can stay there until user clicks "hide ^"... I created the layout with the idea of columns so that structurally we can keep everything organized and in place. But they are all expandable and retractable. Please let me know if there are any other ideas that you would like to see added! I will gladly update based on group input. 

Oh yeah, the idea with the advance search form... what I envision is the user having ability to select what information they would like to see on their search page with what we are providing, and how many results they would like to view from the search result. Maybe on main page (where results are displayed) each row can have a light gray X button that appears on hover giving user the ability